third_party/WebKit/LayoutTests/fast/css/border-shorthand-initialize-longhands-expected.txt

This test ensures border shorthand property initializes longhand properties such as border-top-style and border-right-color.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS div.setAttribute('style', 'border: 1px');div.style.borderTopStyle is 'initial'
PASS div.style.borderTopStyle is 'initial'
PASS div.style.borderRightStyle is 'initial'
PASS div.style.borderBottomStyle is 'initial'
PASS div.style.borderLeftStyle is 'initial'
PASS div.style.borderTopColor is 'initial'
PASS div.style.borderRightColor is 'initial'
PASS div.style.borderBottomColor is 'initial'
PASS div.style.borderLeftColor is 'initial'
PASS div.setAttribute('style', 'border: solid');div.style.borderTopWidth is 'initial'
PASS div.style.borderRightWidth is 'initial'
PASS div.style.borderBottomWidth is 'initial'
PASS div.style.borderLeftWidth is 'initial'
PASS successfullyParsed is true

TEST COMPLETE
